subil: resolved dividebyzero error and layer hit ratio calculation
Corrected complex_cache initializing number of hits and reqs from
ints to floats so that hit ratio can be calculated with floating
point division (rather than integer division, which will always
result in 0).

Also resolved key error that was occuring when extracting layer
information because it was trying to access a key from the item
in the total_trace.json that does not exist. Changed it to directly
at the uri variable which was initialized before and contains the
information we need.