Revert prefetch plugin to what it was
I started working on modifying the prefetch plugin thinking that the
implementation was incorrect. That opinion hasn't changed. But I created
a new file with the old prefetch logic in a file called `prefetch_old.py`
while my WIP changes are in `prefetch.py`. `prefetch_old.py` is what is
currently set as the plugin in `analysis.py`.

The problem I have with the old prefetch plugin code is that: though
Anwar's paper describes using some sort of layer id to identify and add
and remove layers into and out of the prefetch cache, the traces
have nothing suitable for this purpose. ML seems to have used the
layer size as the layer id to distinguish between the different layers.

But until I figure out a better way, we'll just use the prefetch code
as it is.