Ran simulation for LRU cache for portions of traces
I ran the LRU cache simulation (with the code in
cache_usr_repo_layer.py) for 10%, 25%, 50%, 75%, 100% of the total trace
(you can find those in the same dir where total_trace.json is).
The simulation was run with caches of 5%, 10%, 15%, 20%, 30% the size
of the number of unique layers (as obtained by counting the unique
layers in the GET requests) in each of those trace portions.

Next, will be working on getting the same data for Anwar's prefetch.